Cast to wider value first

+++ BrillouinAcquisition/src/Devices/pvcamera.cpp
@@ -297,7 +297,7 @@ void PVCamera::preparePreview() {
 		delete[] m_buffer;
 		m_buffer = nullptr;
 	}
-	int bufSize = circBufferFrames * m_bufferSize / sizeof(PVCam::uns16);
+	int bufSize = (int)circBufferFrames * m_bufferSize / sizeof(PVCam::uns16);
 	m_buffer = new (std::nothrow) PVCam::uns16[bufSize];
 
 	// Start acquisition
